# docs/swagger.yaml
definitions:
...
  pages.Page:
    properties:
      _id:
        type: string
      base:
        $ref: '#/definitions/support.Base'
      code:
        type: string
      content:
        type: string
      meta:
        type: string
      name:
        type: string
      title:
        type: string
    type: object
  support.Base:
    properties:
      created_At:
        type: string
      delete_At:
        type: string
      updated_At:
        type: string
    type: object
  support.HttpError:
    properties:
      code:
        type: integer
      errors:
        type: string
      message:
        type: string
      results:
        type: boolean
    type: object
  tags.Tag:
    properties:
      _id:
        type: string
      base:
        $ref: '#/definitions/support.Base'
      code:
        type: string
      description:
        type: string
      name:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/pages:
    delete:
      consumes:
      - application/json
      description: Delte item
      parameters:
      - description: code
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pages.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Delte a page
      tags:
      - pages
    get:
      consumes:
      - application/json
      description: Get item
      parameters:
      - description: code
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pages.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Get a page
      tags:
      - pages
    post:
      consumes:
      - application/json
      description: Create a new page item
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pages.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Create a page
      tags:
      - pages
    put:
      consumes:
      - application/json
      description: Update an new page item
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pages.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Update an page
      tags:
      - pages
  /api/tags:
    delete:
      consumes:
      - application/json
      description: Delte item
      parameters:
      - description: code
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tags.Tag'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Delte a tag
      tags:
      - tags
    get:
      consumes:
      - application/json
      description: Get item
      parameters:
      - description: code
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tags.Tag'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Get a tag
      tags:
      - tags
    put:
      consumes:
      - application/json
      description: Update an new tag item
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tags.Tag'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/support.HttpError'
      summary: Update an tag
      tags:
      - tags
...
swagger: "2.0"